+++ src/ezzmq/socket.clj
@@ -47,10 +47,45 @@
   (contains? (set (map socket-type-lookup socket-type-kws)) socket-type))
 
 (defn socket
-  [socket-type-kw & [{:keys [bind connect subscribe
+  "Creates a socket and optionally does some things that you typically do with
+   a socket after creating it (see options below).
+
+   `socket-type-kw` must be one of:
+   - :pair
+   - :pub
+   - :sub
+   - :req
+   - :rep
+   - :xreq
+   - :xrep
+   - :dealer
+   - :router
+   - :xpub
+   - :xsub
+   - :pull
+   - :push
+
+   Options:
+   - :identity sets the identity of the socket. The value can be a string or a
+     byte array.
+   - :bind/:connect specify the location on which to bind or connect the socket.
+     The value must be a connection string like \"tcp://*:12345\",
+     \"inproc://something\", etc.
+   - :subscribe specifies one or more subscription topics when creating a SUB
+     socket. By default, the socket will be subscribed to \"\", i.e. all
+     messages. You can override this behavior by using this option to subscribe
+     to specific topics. The value can be a string, a byte array, or a
+     collection of strings or byte arrays.
+   - :send-hwm/:receive-hwm set the send and receive high water mark (HWM), in
+     in milliseconds.
+   - :linger sets the linger period, in milliseconds."
+  [socket-type-kw & [{:keys [identity
+                             bind connect subscribe
                              send-hwm receive-hwm linger]}]]
   (let [socket-type (socket-type-lookup socket-type-kw)
         socket      (create-socket ctx/*context* socket-type)]
+    (when identity
+      (.setIdentity socket (util/as-byte-array identity)))
     (when bind
       (let [bindings (if (coll? bind) bind [bind])]
         (doseq [b bindings]
@@ -61,9 +96,8 @@
           (.connect socket c))))
     (when (socket-type= :sub socket-type)
       (let [topics (if (coll? subscribe) subscribe [(or subscribe "")])]
-        (doseq [t topics]
-          (let [topic (if (string? t) (.getBytes t) t)]
-            (.subscribe socket topic)))))
+        (doseq [topic topics]
+          (.subscribe socket (util/as-byte-array topic)))))
     (when send-hwm (.setSndHWM socket send-hwm))
     (when receive-hwm (.setRcvHWM socket receive-hwm))
     (when linger (.setLinger socket linger))