+++ src/ezzmq/util.clj
@@ -4,3 +4,14 @@
   [msg & fmt-args]
   (throw (Exception. (apply format msg fmt-args))))
 
+(defn as-byte-array
+  [x]
+  (condp = (class x)
+    (Class/forName "[B")
+    x
+
+    String
+    (.getBytes x)
+
+    (errfmt "Can't convert %s (%s) into a byte array." x (class x))))
+